core/pyproject.toml
[project]
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.115.14",
    "orjson>=3.11.5",
    "passlib>=1.7.4",
    "psutil>=7.2.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.21",
    "sqlmodel>=0.0.30",
    "uvicorn>=0.35.0",
]
name = "core"
readme = "README.md"
requires-python = ">=3.13"
version = "0.0.0"

[build-system]
build-backend = "uv_build"
requires = ["uv_build>=0.9.4,<0.10.0"]

[dependency-groups]
dev = [
    "types-aioboto3>=15.5.0",
    "types-passlib>=1.7.7.20250602",
    "types-pillow>=10.2.0",
    "types-psutil>=7.1.3.20251211",
]
